refactor(fee-calculator): rename intent_fee_calculator to fee_calculator
- Rename src/lib/intent_fee_calculator/ to src/lib/fee_calculator/
- Update package name in crate Cargo.toml
- Update workspace member path and dependency in root Cargo.toml
- Update backend Cargo.toml dependency
- Update import in utils.rs

Phase 1 of fee calculator refactor plan.
