+++ src/lib/fee_calculator/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "intent_fee_calculator"
+name = "fee_calculator"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
