+++ src/lib/fee_calculator/src/flow_detector.rs
